CHANGES: 1) Introduced lattice_models_interafce.h where I put the abstract interface for both binomial and trinomial models 2) Further seperated models into lattice_model_binomial.h an lattice_model_trinomial.h modules. Module lattice_model.h contains includes for both lattice_model_binomial.h and lattice_trinomial.h 3) Started on Black-Karasinski model
TODO:
 - Finish Black-Karasinski model calibration with forward induction
 algo.
 - Look at the backward induction algo for trinomial models.
 In particular,those with revertinig feature.
 - Try simplify MeanRevertingParams structure - consider passing
 OptionData structure to it

+++ lattice_calibrator_ir_t.h
@@ -955,7 +955,7 @@ void testHWSanityCheck() {
 		fixingDatesSet.emplace(today + date_duration(daysInhalfYear*t));
 	}
 
-	float daysInYear{ 365.0 };
+	double daysInYear{ 365.0 };
 	std::vector<double> timeDeltas(fixingDates.size() - 1);
 	for (auto i = 0; i < timeDeltas.size(); ++i) {
 		timeDeltas[i] = ((fixingDates[i + 1] - fixingDates[i]).days() / daysInYear);
